refactor: Move email channel database classes.
