+++ alert-database-job/build.gradle
@@ -1,0 +1,23 @@
+ext.moduleName = 'com.synopsys.integration.alert.database.job'
+
+dependencies {
+    implementation platform(project(':alert-platform'))
+
+    implementation project(':alert-common')
+    implementation project(':api-common-model')
+    implementation project(':api-descriptor')
+
+    implementation project(':api-provider')
+    implementation project(':alert-database')
+    implementation project(':channel-email')
+
+    api 'org.springframework.data:spring-data-jpa'
+    implementation 'jakarta.persistence:jakarta.persistence-api'
+    implementation 'com.zaxxer:HikariCP'
+    implementation 'org.springframework.boot:spring-boot-autoconfigure'
+    implementation 'org.springframework.security:spring-security-core'
+    implementation 'org.hibernate:hibernate-core'
+
+    testImplementation project(':test-common')
+    testImplementation 'com.synopsys.integration:blackduck-common', rootProject.ext.blackduckCommonExcludes
+}
