+++ alert-database-job/src/main/java/com/synopsys/integration/alert/database/api/StaticJobAccessor.java
