+++ channel-email/src/main/java/com/synopsys/integration/alert/channel/email/database/accessor/DefaultEmailJobDetailsAccessor.java
@@ -5,7 +5,7 @@
  *
  * Use subject to the terms and conditions of the Synopsys End User Software License and Maintenance Agreement. All rights reserved worldwide.
  */
-package com.synopsys.integration.alert.database.job.email;
+package com.synopsys.integration.alert.channel.email.database.accessor;
 
 import java.util.List;
 import java.util.Optional;
@@ -17,11 +17,13 @@ import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.synopsys.integration.alert.channel.email.database.job.EmailJobDetailsEntity;
+import com.synopsys.integration.alert.channel.email.database.job.EmailJobDetailsRepository;
+import com.synopsys.integration.alert.channel.email.database.job.additional.EmailJobAdditionalEmailAddressEntity;
+import com.synopsys.integration.alert.channel.email.database.job.additional.EmailJobAdditionalEmailAddressRepository;
 import com.synopsys.integration.alert.common.persistence.accessor.EmailJobDetailsAccessor;
 import com.synopsys.integration.alert.common.persistence.model.job.details.DistributionJobDetailsModel;
 import com.synopsys.integration.alert.common.persistence.model.job.details.EmailJobDetailsModel;
-import com.synopsys.integration.alert.database.job.email.additional.EmailJobAdditionalEmailAddressEntity;
-import com.synopsys.integration.alert.database.job.email.additional.EmailJobAdditionalEmailAddressRepository;
 import com.synopsys.integration.alert.descriptor.api.EmailChannelKey;
 import com.synopsys.integration.alert.descriptor.api.model.DescriptorKey;
