+++ channel-email/src/main/java/com/synopsys/integration/alert/channel/email/database/configuration/EmailConfigurationRepository.java
@@ -5,11 +5,15 @@
  *
  * Use subject to the terms and conditions of the Synopsys End User Software License and Maintenance Agreement. All rights reserved worldwide.
  */
-package com.synopsys.integration.alert.database.job.email;
+package com.synopsys.integration.alert.channel.email.database.configuration;
 
+import java.util.Optional;
 import java.util.UUID;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EmailJobDetailsRepository extends JpaRepository<EmailJobDetailsEntity, UUID> {
+public interface EmailConfigurationRepository extends JpaRepository<EmailConfigurationEntity, UUID> {
+
+    Optional<EmailConfigurationEntity> findByName(String name);
+
 }
