+++ channel-email/src/main/java/com/synopsys/integration/alert/channel/email/database/configuration/properties/EmailConfigurationPropertiesRepository.java
@@ -5,7 +5,7 @@
  *
  * Use subject to the terms and conditions of the Synopsys End User Software License and Maintenance Agreement. All rights reserved worldwide.
  */
-package com.synopsys.integration.alert.database.email.properties;
+package com.synopsys.integration.alert.channel.email.database.configuration.properties;
 
 import java.util.UUID;
