+++ channel-email/src/main/java/com/synopsys/integration/alert/channel/email/database/job/EmailJobDetailsRepository.java
@@ -5,15 +5,11 @@
  *
  * Use subject to the terms and conditions of the Synopsys End User Software License and Maintenance Agreement. All rights reserved worldwide.
  */
-package com.synopsys.integration.alert.database.email;
+package com.synopsys.integration.alert.channel.email.database.job;
 
-import java.util.Optional;
 import java.util.UUID;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EmailConfigurationRepository extends JpaRepository<EmailConfigurationEntity, UUID> {
-
-    Optional<EmailConfigurationEntity> findByName(String name);
-
+public interface EmailJobDetailsRepository extends JpaRepository<EmailJobDetailsEntity, UUID> {
 }
