+++ channel-email/src/main/java/com/synopsys/integration/alert/channel/email/database/job/additional/EmailJobAdditionalEmailAddressEntity.java
@@ -5,7 +5,7 @@
  *
  * Use subject to the terms and conditions of the Synopsys End User Software License and Maintenance Agreement. All rights reserved worldwide.
  */
-package com.synopsys.integration.alert.database.job.email.additional;
+package com.synopsys.integration.alert.channel.email.database.job.additional;
 
 import java.util.UUID;
