+++ settings.gradle
@@ -2,6 +2,7 @@ rootProject.name = 'blackduck-alert'
 include 'azure-boards-common'
 include 'alert-common'
 include 'alert-database'
+include 'alert-database-job'
 include 'alert-platform'
 include 'api-channel'
 include 'api-channel-issue-tracker'
